This merge updates Clacc to upstream LLVM main at abf041632821 (local
CommitDate 2022/08/28), bringing in the following commits:

* 51bda3a0e7ac: This commit changes uses of `std::vector` to
  `llvm::SmallVector`, including in the `TgtPtrs` member of
  `PrivateArgumentManagerTy` in
  `openmp/libomptarget/src/omptarget.cpp`.  For the sake of OMPT
  support, Clacc renames `TgtPtrs` to `FirstPrivateTgtInfo` and
  changes its vector element type.  This merge keeps Clacc's version
  except updates `std::vector` to `SmallVector`.
* 540a13652fda: This commit replaces uses of `dlopen` and `dlsym` with
  LLVM's `loadPermanentLibrary` and `getSymbolAddress`.  This merge:
    * Replaces a `dlsym` use for `__tgt_rtl_get_device_type` added by
      Clacc to `openmp/libomptarget/src/rtl.cpp`.
    * Replaces a `dlsym` use for `acc2omp_set_omp_default_device`
      added by Clacc to `openmp/libomptarget/src/rtl.cpp`.
    * Adjusts various `openmp/libacc2omp` comments not to imply
      `dlopen` and `dlsym` are used given that LLVM's implementation
      of `loadPermanentLibrary` and `getSymbolAddress` might be
      different on some operating systems.
    * Does not yet convert `openmp/libacc2omp` to an LLVM library and
      replace uses of `dlopen` and `dlsym` there.  Upstream hasn't
      converted `openmp/runtime` either.
* 992e10a3fce4: This commit fixes check-all to include check-openmp.
  This merge adjusts Clacc's CI config to remove the workaround of
  specifying check-openmp separately.
* b5f8d42efe3e, 5512f398a039e: These commits deprecate various
  functions in `llvm/include/llvm/ADT/Optional.h`, producing warnings
  and fixit suggestions.  This merge updates Clacc's uses by applying
  the fixit suggestions.
* 82e893c47c77: This commit introduces
  `clang/test/Frontend/sarif-diagnostics.cpp`, a test that passes at
  abf041632821 but somehow fails after being merged into Clacc.  The
  problem is differences in indices appearing in the diagnostic output
  under test.  We assume that whatever in Clacc is causing those
  indices to change is benign and that the test is just fragile.
  Indeed, this merge cherry-picks upstream's later b345be177d03, which
  also notes that these indices are fragile and adjusts the test not
  to strictly match them.  As a result, it now passes.
* Something has fixed the OpenACC V&V test `parallel_loop_worker.c` on
  ExCL's explorer (AMD GPU).  It was broken by the previous merge
  (374be6b62bbc).
* Various commits with which this merge resolves contextual conflicts.

This merge updates Clacc's CI config's `LIT_XFAIL`, `LIT_XFAIL_NOT`,
and `LIT_FILTER_OUT` lists based on tests results at abf041632821.

+++ lld/MachO/UnwindInfoSection.cpp
@@ -82,18 +82,10 @@ using namespace lld::macho;
 // advantage, achieving a 3-order-of-magnitude reduction in the
 // number of entries.
 //
-// * The __TEXT,__unwind_info format can accommodate up to 127 unique
-// encodings for the space-efficient compressed format. In practice,
-// fewer than a dozen unique encodings are used by C++ programs of
-// all sizes. Therefore, we don't even bother implementing the regular
-// non-compressed format. Time will tell if anyone in the field ever
-// overflows the 127-encodings limit.
-//
 // Refer to the definition of unwind_info_section_header in
 // compact_unwind_encoding.h for an overview of the format we are encoding
 // here.
 
-// TODO(gkm): prune __eh_frame entries superseded by __unwind_info, PR50410
 // TODO(gkm): how do we align the 2nd-level pages?
 
 // The offsets of various fields in the on-disk representation of each compact
@@ -401,8 +393,12 @@ static bool canFoldEncoding(compact_unwind_encoding_t encoding) {
   // of the unwind info's unwind address, two functions that have identical
   // unwind info can't be folded if it's using this encoding since both
   // entries need unique addresses.
-  static_assert(UNWIND_X86_64_MODE_MASK == UNWIND_X86_MODE_MASK, "");
-  static_assert(UNWIND_X86_64_MODE_STACK_IND == UNWIND_X86_MODE_STACK_IND, "");
+  static_assert(static_cast<uint32_t>(UNWIND_X86_64_MODE_MASK) ==
+                    static_cast<uint32_t>(UNWIND_X86_MODE_MASK),
+                "");
+  static_assert(static_cast<uint32_t>(UNWIND_X86_64_MODE_STACK_IND) ==
+                    static_cast<uint32_t>(UNWIND_X86_MODE_STACK_IND),
+                "");
   if ((target->cpuType == CPU_TYPE_X86_64 || target->cpuType == CPU_TYPE_X86) &&
       (encoding & UNWIND_X86_64_MODE_MASK) == UNWIND_X86_64_MODE_STACK_IND) {
     // FIXME: Consider passing in the two function addresses and getting
